Properly make pytest fixture plugins
